// users-service/src/domain/enum/leadStatus.enum.ts
export enum LeadStatus {
  OTP_NOT_SENT = 'OTP_NOT_SENT',
  OTP_GENERATED = 'OTP_GENERATED',
  OTP_VERIFIED = 'OTP_VERIFIED',
  OTP_FAILED = 'OTP_FAILED',
  DEDUPE_FAILED = 'DEDUPE_FAILED',
  DEDUPE_SUCCESS = 'DEDUPE_SUCCESS',
  LEAD_CREATED = 'LEAD_CREATED',
  KYC_DONE = 'KYC_DONE',
  SANCTION_DONE = 'SANCTION_DONE',
  LEAD_VERIFIED = 'LEAD_VERIFIED',
  SCORING_DONE = 'SCORING_DONE',
  LEAD_ENHANCED = 'LEAD_ENHANCED',
  LEAD_ONBOARDED = 'LEAD_ONBOARDED',
}